requirements.txt
netmiko
pyyaml